# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "pokemon-api",
    "description": "API Documentation for the Pokemon API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/pokedex/": {
      "post": {
        "tags": [
          "Pokedex"
        ],
        "description": "Create a new entry in the Pokedex",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Pokedex information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "object",
                  "properties": {
                    "english": {
                      "type": "string",
                      "example": "EnglishName"
                    },
                    "japanese": {
                      "type": "string",
                      "example": "JapaneseName"
                    },
                    "chinese": {
                      "type": "string",
                      "example": "ChineseName"
                    },
                    "french": {
                      "type": "string",
                      "example": "FrenchName"
                    }
                  }
                },
                "type": {
                  "type": "array",
                  "example": [
                    "typeid"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "base": {
                  "type": "object",
                  "properties": {
                    "HP": {
                      "type": "string",
                      "example": "Hit Points"
                    },
                    "Attack": {
                      "type": "string",
                      "example": "Attack"
                    },
                    "Defense": {
                      "type": "string",
                      "example": "Defense"
                    },
                    "Sp. Attack": {
                      "type": "string",
                      "example": "Special Attack"
                    },
                    "Sp. Defense": {
                      "type": "string",
                      "example": "Special Defense"
                    },
                    "Speed": {
                      "type": "string",
                      "example": "Speed"
                    }
                  }
                },
                "moves": {
                  "type": "array",
                  "example": [
                    "MovesIds"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "name",
                "type",
                "base",
                "moves"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Pokedex"
        ],
        "description": "Retrieve all entries from the Pokedex.",
        "responses": {
          "302": {
            "description": "Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/pokedex/{id}": {
      "get": {
        "tags": [
          "Pokedex"
        ],
        "description": "Retrieve a specific entry from the Pokedex by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the Pokedex entry"
          }
        ],
        "responses": {
          "302": {
            "description": "Found"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Pokedex"
        ],
        "description": "Update a specific entry in the Pokedex by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the Pokedex entry"
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated information for the Pokedex entry",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "object",
                  "properties": {
                    "english": {
                      "type": "string",
                      "example": "EnglishName"
                    },
                    "japanese": {
                      "type": "string",
                      "example": "JapaneseName"
                    },
                    "chinese": {
                      "type": "string",
                      "example": "ChineseName"
                    },
                    "french": {
                      "type": "string",
                      "example": "FrenchName"
                    }
                  }
                },
                "type": {
                  "type": "array",
                  "example": [
                    "typeid"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "base": {
                  "type": "object",
                  "properties": {
                    "HP": {
                      "type": "string",
                      "example": "Hit Points"
                    },
                    "Attack": {
                      "type": "string",
                      "example": "Attack"
                    },
                    "Defense": {
                      "type": "string",
                      "example": "Defense"
                    },
                    "Sp. Attack": {
                      "type": "string",
                      "example": "Special Attack"
                    },
                    "Sp. Defense": {
                      "type": "string",
                      "example": "Special Defense"
                    },
                    "Speed": {
                      "type": "string",
                      "example": "Speed"
                    }
                  }
                },
                "moves": {
                  "type": "array",
                  "example": [
                    "MovesIds"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "name",
                "type",
                "base",
                "moves"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Pokedex"
        ],
        "description": "Replace a specific entry in the Pokedex by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the Pokedex entry"
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated information for the Pokedex entry",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "object",
                  "properties": {
                    "english": {
                      "type": "string",
                      "example": "EnglishName"
                    },
                    "japanese": {
                      "type": "string",
                      "example": "JapaneseName"
                    },
                    "chinese": {
                      "type": "string",
                      "example": "ChineseName"
                    },
                    "french": {
                      "type": "string",
                      "example": "FrenchName"
                    }
                  }
                },
                "type": {
                  "type": "array",
                  "example": [
                    "typeid"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "base": {
                  "type": "object",
                  "properties": {
                    "HP": {
                      "type": "string",
                      "example": "Hit Points"
                    },
                    "Attack": {
                      "type": "string",
                      "example": "Attack"
                    },
                    "Defense": {
                      "type": "string",
                      "example": "Defense"
                    },
                    "Sp. Attack": {
                      "type": "string",
                      "example": "Special Attack"
                    },
                    "Sp. Defense": {
                      "type": "string",
                      "example": "Special Defense"
                    },
                    "Speed": {
                      "type": "string",
                      "example": "Speed"
                    }
                  }
                },
                "moves": {
                  "type": "array",
                  "example": [
                    "MovesIds"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "name",
                "type",
                "base",
                "moves"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Pokedex"
        ],
        "description": "Delete a specific entry from the Pokedex by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the Pokedex entry"
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
...
}